ST-791: Intial changes to update data required to display on Summary tab
