+++ src/main/java/uk/gov/hmcts/sptribs/notification/NotificationServiceCIC.java
@@ -3,6 +3,8 @@ package uk.gov.hmcts.sptribs.notification;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import uk.gov.hmcts.sptribs.ciccase.model.NotificationResponse;
+import uk.gov.hmcts.sptribs.ciccase.model.NotificationType;
 import uk.gov.hmcts.sptribs.common.config.EmailTemplatesConfigCIC;
 import uk.gov.hmcts.sptribs.notification.exception.NotificationException;
 import uk.gov.hmcts.sptribs.notification.model.NotificationRequest;
@@ -10,8 +12,11 @@ import uk.gov.service.notify.NotificationClient;
 import uk.gov.service.notify.NotificationClientException;
 import uk.gov.service.notify.SendEmailResponse;
 
+import java.time.LocalDateTime;
 import java.util.Map;
 import java.util.UUID;
+
+import static java.util.Objects.nonNull;
 
 @Service
 @Slf4j
@@ -25,7 +30,7 @@ public class NotificationServiceCIC {
     @Autowired
     private EmailTemplatesConfigCIC emailTemplatesConfig;
 
-    public SendEmailResponse sendEmail() {
+    public NotificationResponse sendEmail() {
         SendEmailResponse sendEmailResponse = null;
         String destinationAddress = notificationRequest.getDestinationAddress();
         EmailTemplateName template = notificationRequest.getTemplate();
@@ -59,11 +64,24 @@ public class NotificationServiceCIC {
             );
             throw new NotificationException(notificationClientException);
         }
-        return sendEmailResponse;
+        return getNotificationResponse(sendEmailResponse);
     }
 
     public void setNotificationRequest(NotificationRequest notificationRequest) {
         this.notificationRequest = notificationRequest;
+    }
+
+    private NotificationResponse getNotificationResponse(final SendEmailResponse sendEmailResponse) {
+        String clientReference = nonNull(sendEmailResponse) ? sendEmailResponse.getReference().orElse(null) : null;
+        UUID id = nonNull(sendEmailResponse) ? sendEmailResponse.getNotificationId() : null;
+
+        return NotificationResponse.builder()
+            .id(id.toString())
+            .client_reference(clientReference)
+            .notificationType(NotificationType.EMAIL)
+            .updatedAtTime(LocalDateTime.now())
+            .createdAtTime(LocalDateTime.now())
+            .build();
     }
 
 }
